updates to flitering and imports
I think we are using the filtering function wrong. according to his notes he is looking for something like the comments on line 111-115. I also imported matplotlib.pyplot as plt this is very helpful for doing tedious calculations. one other thing that i noticed is we are not using a 2nd matrix for check/ observations. that might be something we want to discuss. I will make more changes tonight as comments if you would like we can make a separate branch for my implementations